feat(ShortcutMenu): Move package
