+++ packages/code-editor/src/ShortcutMenu/index.ts
